requirements.txt
absl-py==0.7.1
astor==0.8.0
Click==7.0
exception==0.1.0
Flask==1.1.1
gast==0.2.2
grpcio==1.23.0
h5py==2.9.0
itsdangerous==1.1.0
Jinja2==2.10.1
Keras-Applications==1.0.8
Keras-Preprocessing==1.1.0
MarkupSafe==1.1.1
numpy==1.17.0
pg8000==1.13.2
protobuf==3.9.1
psycopg2==2.8.3
py==1.8.0
requests==2.22.0
scramp==1.1.0
six==1.12.0
SQLAlchemy==1.2.19
tensorboard==1.14.0
tensorflow==1.14.0
tensorflow-estimator==1.14.0
termcolor==1.1.0
testing.common.database==2.0.3
testing.postgresql==1.3.0
urllib3==1.25.3
Werkzeug==0.15.5
wrapt==1.11.2
